requirements.txt
accelerate==0.30.0
aiohttp==3.9.5
aiokafka==0.10.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.98
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
colorama==0.4.6
dill==0.3.8
docker==7.0.0
docutils==0.16
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
google-pasta==0.2.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==6.11.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nextcord==2.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathos==0.3.2
platformdirs==4.2.1
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-env==1.0.0
pytz==2024.1
PyYAML==6.0.1
ratelimit==2.2.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.1
safetensors==0.4.3
sagemaker==2.218.1
schema==0.7.7
six==1.16.0
smdebug-rulesconfig==1.0.1
sympy==1.12
tblib==3.0.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
